Notebooks/Lakshit/GeminiChatbot/requirements.txt
pandas==2.2.2
numpy==1.26.4
sentence-transformers==2.6.1
faiss-cpu==1.7.4 ; sys_platform != 'darwin'
google-generativeai==0.4.1
chainlit==1.3.2
python-dotenv